ref: move immutable-array-fns to /shared directory
frontend seems to need it too :(
